don't instantiate argument defaults in the function declaration
doing so means they get instantiated on import, and potentially reused
across multiple calls to a function.

also, if you're running as a user that doesn't have a home directory,
that means that the `Hyphenator('en_US')` call fails with a PermissionError
before the caller has a chance to pass their own in.